Place thread local items into its own class
