+++ core/src/main/resources/filodb-defaults.conf
@@ -218,15 +218,15 @@ filodb {
     # Number of time series to operate on at one time. Reduce if there is much less memory available
     num-partitions-per-cass-write = 10000
 
-    # This block memory is used for overflow of write buffers and for storing encoded chunks
+    # This block memory is used for overflow of write buffers and for storing encoded downsample chunks
     # Configure to NumPartitionsPerCassWrite * (EstChunkSize * EstNumChunksPerPartition) + buffer
-    # 10000 * (5KB * 3) + buffer
-    off-heap-block-memory-size = 500 MB
+    # 10000 * (1KB per col * 6 cols * 3 chunks) + buffer
+    off-heap-block-memory-size = 400 MB
 
     # Used for part keys, ODPed raw chunks, write buffers for downsampled data
     # Configure to NumPartitionsPerCassWrite * (EstRawChunkSize * EstNumRawChunksPerPartition + PartKeySize + BytesPerDownsampledRow * NumDownsampledRowsPerPart) + buffer
-    # 10000 * (1KB * 2 + 500 bytes + 20b * 360) + buffer
-    off-heap-native-memory-size = 500 MB
+    # 10000 * (1KB * 2 cols * 8 chunks + 500 bytes + 20B * 360 samples) + buffer
+    off-heap-native-memory-size = 400 MB
 
     # Amount of time to wait for a Cassandra write to finish before proceeding to next batch of partitions
     cassandra-write-timeout = 10.minutes
